fix(components): file uploader and input styles
