+++ packages/carbon-web-components/src/components/input/input.scss
@@ -9,7 +9,7 @@ $css--plex: true !default;
 
 @use '@carbon/styles/scss/config' as *;
 @use '@carbon/styles/scss/spacing' as *;
-@use '@carbon/styles/scss/components/text-input/text-input';
+@use '@carbon/styles/scss/components/text-input/index';
 
 :host(#{$prefix}-input) {
   // the base text-input class has a 100% so we need to propagate it up to the host element
